# backend/api/gen/go/app/service/v1/i_finance_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: app/service/v1/i_finance.proto

package apppb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-erp/api/gen/go/finance/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPayableServiceGet = "/app.service.v1.PayableService/Get"
const OperationPayableServiceList = "/app.service.v1.PayableService/List"

type PayableServiceHTTPServer interface {
	// Get 查询应付单详情
	Get(context.Context, *v11.GetPayableRequest) (*v11.Payable, error)
	// List 查询应付单列表
	List(context.Context, *v1.PagingRequest) (*v11.ListPayableResponse, error)
}

func RegisterPayableServiceHTTPServer(s *http.Server, srv PayableServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/app/v1/payables", _PayableService_List1_HTTP_Handler(srv))
	r.GET("/app/v1/payables/{id}", _PayableService_Get1_HTTP_Handler(srv))
}

func _PayableService_List1_HTTP_Handler(srv PayableServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayableServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPayableResponse)
		return ctx.Result(200, reply)
	}
}

func _PayableService_Get1_HTTP_Handler(srv PayableServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPayableRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayableServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPayableRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payable)
		return ctx.Result(200, reply)
	}
}

type PayableServiceHTTPClient interface {
	// Get 查询应付单详情
	Get(ctx context.Context, req *v11.GetPayableRequest, opts ...http.CallOption) (rsp *v11.Payable, err error)
	// List 查询应付单列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPayableResponse, err error)
}

type PayableServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPayableServiceHTTPClient(client *http.Client) PayableServiceHTTPClient {
	return &PayableServiceHTTPClientImpl{client}
}

// Get 查询应付单详情
func (c *PayableServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPayableRequest, opts ...http.CallOption) (*v11.Payable, error) {
	var out v11.Payable
	pattern := "/app/v1/payables/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayableServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询应付单列表
func (c *PayableServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPayableResponse, error) {
	var out v11.ListPayableResponse
	pattern := "/app/v1/payables"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayableServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

const OperationPaymentServiceGet = "/app.service.v1.PaymentService/Get"
const OperationPaymentServiceList = "/app.service.v1.PaymentService/List"

type PaymentServiceHTTPServer interface {
	// Get 查询付款详情
	Get(context.Context, *v11.GetPaymentRequest) (*v11.Payment, error)
	// List 查询付款列表
	List(context.Context, *v1.PagingRequest) (*v11.ListPaymentResponse, error)
}

func RegisterPaymentServiceHTTPServer(s *http.Server, srv PaymentServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/app/v1/payments", _PaymentService_List2_HTTP_Handler(srv))
	r.GET("/app/v1/payments/{id}", _PaymentService_Get2_HTTP_Handler(srv))
}

func _PaymentService_List2_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPaymentResponse)
		return ctx.Result(200, reply)
	}
}

func _PaymentService_Get2_HTTP_Handler(srv PaymentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPaymentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPaymentServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.Payment)
		return ctx.Result(200, reply)
	}
}

type PaymentServiceHTTPClient interface {
	// Get 查询付款详情
	Get(ctx context.Context, req *v11.GetPaymentRequest, opts ...http.CallOption) (rsp *v11.Payment, err error)
	// List 查询付款列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPaymentResponse, err error)
}

type PaymentServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPaymentServiceHTTPClient(client *http.Client) PaymentServiceHTTPClient {
	return &PaymentServiceHTTPClientImpl{client}
}

// Get 查询付款详情
func (c *PaymentServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPaymentRequest, opts ...http.CallOption) (*v11.Payment, error) {
	var out v11.Payment
	pattern := "/app/v1/payments/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询付款列表
func (c *PaymentServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPaymentResponse, error) {
	var out v11.ListPaymentResponse
	pattern := "/app/v1/payments"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPaymentServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}