# backend/api/gen/go/app/service/v1/i_procurement_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: app/service/v1/i_procurement.proto

package apppb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-erp/api/gen/go/procurement/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPurchaseOrderServiceGet = "/app.service.v1.PurchaseOrderService/Get"
const OperationPurchaseOrderServiceList = "/app.service.v1.PurchaseOrderService/List"

type PurchaseOrderServiceHTTPServer interface {
	// Get 查询采购单详情（含明细）
	Get(context.Context, *v11.GetPurchaseOrderRequest) (*v11.PurchaseOrder, error)
	// List 查询采购单列表
	List(context.Context, *v1.PagingRequest) (*v11.ListPurchaseOrderResponse, error)
}

func RegisterPurchaseOrderServiceHTTPServer(s *http.Server, srv PurchaseOrderServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/app/v1/purchase-orders", _PurchaseOrderService_List6_HTTP_Handler(srv))
	r.GET("/app/v1/purchase-orders/{id}", _PurchaseOrderService_Get6_HTTP_Handler(srv))
}

func _PurchaseOrderService_List6_HTTP_Handler(srv PurchaseOrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPurchaseOrderServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListPurchaseOrderResponse)
		return ctx.Result(200, reply)
	}
}

func _PurchaseOrderService_Get6_HTTP_Handler(srv PurchaseOrderServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetPurchaseOrderRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPurchaseOrderServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetPurchaseOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.PurchaseOrder)
		return ctx.Result(200, reply)
	}
}

type PurchaseOrderServiceHTTPClient interface {
	// Get 查询采购单详情（含明细）
	Get(ctx context.Context, req *v11.GetPurchaseOrderRequest, opts ...http.CallOption) (rsp *v11.PurchaseOrder, err error)
	// List 查询采购单列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListPurchaseOrderResponse, err error)
}

type PurchaseOrderServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewPurchaseOrderServiceHTTPClient(client *http.Client) PurchaseOrderServiceHTTPClient {
	return &PurchaseOrderServiceHTTPClientImpl{client}
}

// Get 查询采购单详情（含明细）
func (c *PurchaseOrderServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetPurchaseOrderRequest, opts ...http.CallOption) (*v11.PurchaseOrder, error) {
	var out v11.PurchaseOrder
	pattern := "/app/v1/purchase-orders/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPurchaseOrderServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询采购单列表
func (c *PurchaseOrderServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListPurchaseOrderResponse, error) {
	var out v11.ListPurchaseOrderResponse
	pattern := "/app/v1/purchase-orders"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPurchaseOrderServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}